Add osbash helper scripts
This changeset adds two simple helper scripts.

base_fixups.sh is run on every basedisk while it is built. The script
can be used to fix whatever is necessary to make the basedisk work.
Right now, all it does is remove 70-persistent-net.rules which (if left
untouched) would result in a network interface order that would differ
from Vagrant's.

The second script is a trivial debug helper that does nothing but wait
for the removal of a file it creates in /tmp. When the file is removed,
execution continues.

Partial-Bug: 1312764
Implements: blueprint openstack-training-labs
Change-Id: Id75a89ced5dcf132e91e7426df15436de585a203