Set up keystone databases
This script sets up MySQL databases for keystone, glance, neutron, nova,
and cinder.

The script is just a sample implementation. Depending on how the
training-guides are structured, it may be better to have a function that
can be called in service-specific scripts to create the database they
need.

Partial-Bug: 1312764
Implements: blueprint openstack-training-labs
Change-Id: Iaeca8e637c53a6b91d12915c8855b5b53639afd1