Add localrc configuration file
The localrc file is read by osbash/wbatch and a few guest-side scripts.
It can be used to override variables used by the scripts.

Most notably, this is the place to select the Linux distribution to use
for the base disk or the VM cluster.

All variables have default values; localrc may be left empty.

Partial-Bug: 1312764
Implements: blueprint openstack-training-labs
Change-Id: I38e305ac40eda062bda4433178b7aed3c9e8e694