Don't look for ISO image when only doing wbatch
The current code looks for the installation ISO (and downloads it if
the ISO is missing) even if the user requested that only Windows batch
scripts be written.

With this patch, osbash calls find_install-iso only if it was asked to
build a basedisk on the local machine.

Partial-Bug: 1312764
Implements: blueprint openstack-training-labs
Change-Id: I7af15dcd4fa99671a85a924b23f434303049fcd2