labs: unified login/password handling
The labs scripts need two sets of login/password for each service:
one for the (MySQL) database used by the service, and one for
authentication with keystone.

For the labs scripts login names and passwords are simple functions
of the service names: for the first set (database), we have
${service}User/${service}Pass, for the second set (keystone auth)
${service}/${service}_pass.

This changeset adds two functions for producing the second set of
credentials and renames the existing functions (as well as related
variables) for producing the first set.

This prevents us from having to store login names and passwords in the
service scripts. It also keeps those things in one location where they
can be changed easily should we ever want to do so.

Implements: blueprint openstack-training-labs
Change-Id: I730eab475bee843767984b6c64bab88a94c0156e