labs: osbash CLI default changes
This patch changes the default behavior of osbash in order to improve
useability and better protect user data:

If the user asks for a cluster, the script no longer asks whether to use
a matching base disk (the prompt was added before the basedisk command
that allows re-building a base disk if needed).

If an existing base disk is used, VMs already attached to that disk are
no longer removed wholesale. They are removed only if and when they are
about to be replaced by a new VM of the same name. Among other things,
this will allow us to rebuild some nodes without affecting others.

Change-Id: I6e52c521b9a5f61fa42b644a60c49bf20c2e7cc3
Implements: blueprint openstack-training-labs