labs: fix multiattach disks
When using VBoxMange storageattach with the "--mtype multiattach"
option, the command fails if another VM is already attached (via
multiattach) and running at the time of the call. If the option
is left out, the new VM has its disk multiattached as desired.

Rather than trying to find out when to use the option, this patch uses
modifyhd to register a disk as type multiattach. This call does not fail
if made several times for the same disk, and all subsequent
storageattach calls for that disk have the multiattach option
automatically set.

Change-Id: I5306b855f0736d3e6a00761c514b4a031f167ead
Implements: blueprint openstack-training-labs