labs: add init_node command
This changeset adds a new init_node command for config/scripts.* and
uses it to merge scripts.nodeinit_osbash and scripts.nodeinit_vagrant
into the node specific config/scripts.*.

This results in some redundancy, but it (hopefully) makes the script
configuration easier to follow. It also makes the initial creation and
configuration of the VM just another step that can be picked or omitted
from script configuration.

Implements: blueprint openstack-training-labs
Change-Id: Ieadfe2b9540710556100051723e3e6f08855efa3