labs: add option for VM gui type
This changeset adds an environment variable and an option that can be
used to change the type of VirtualBox VM GUI used while building the
nodes.

Valid values are gui (default), sdl (minimal gui), and headless.

The resulting node VMs can then be started with or without GUI as
before.

Change-Id: If12f4ba6dc455972b6332318ed1261e2c9de8c3f