Adds neutron snapshot to the controller node
Snapshot 'neutron-controller_installed' will contain the following
packages

    1. neutron-server
    2. neutron-plugin-ml2

This snapshot installs and configures neutron. Also configures
keystone to authenticate with neutron.

This snapshot can be used to obtain stable running version of
controller node and can be used to revert back to this point in case
the students break the cluster.

Change-Id: I6df01c67ea8dd1a497f80629704343e732d8d611