labs: wait for namespaces and interfaces
When setting up neutron router and DHCP, the time until namespaces and
interfaces are ready varies a lot (from less than a second to five
seconds and more), even on the same host system. If it takes longer, we
proceed too early which leaves the cluster misconfigured. In that case,
nova instances fail to get a DHCP address.

This patch makes sure the script waits until the interfaces are ready.

Change-Id: I346a098062d984956cf496e89f9ada2a38df0569