labs: set lower MTU for bridge to instance VM
The MTU on the bridge to the instance VMs needs to be lower than 1500 to
compensate for the GRE overhead. With an MTU of 1500, one thing that
fails is ssh authentication (into the instance VM).

This patch does not help for the CirrOS 0.3.2 image which we currently
use, because it is not able to pick up MTU values via DHCP.

Change-Id: I2a04b62beef6120e811f00472aa7395ab3d683dc
Co-Authored-By: Pranav Salunke <dguitarbite@gmail.com>