labs: CirrOS image: check md5sum; URL in config/openstack
Copy the CirrOS image to the base disk even if the IMG_DIR is shared
with the host computer (the image in IMG_DIR remains as a cache).

Check md5sum for CirrOS image after copying it to the base disk.

CIRROS_URL is now stored in config/openstack to allow CirrOS version
upates in one location.

Change-Id: If0cf35d3806e3478875c5cceeaceca683fe1a58e